feat: call with instant hash
