+++ src/core/transactions.ts
@@ -24,7 +24,6 @@ interface PoolEvent {
 
 export interface TxSerialized {
   type:
-    | "PreflightTx"
     | "TxAddLiquidity"
     | "TxRemoveLiquidity"
     | "TxSwap"
@@ -106,8 +105,6 @@ export class TxSerializer {
           return new TxAddBot(params);
         case "TxRemoveBot":
           return new TxRemoveBot(params);
-        case "PreflightTx":
-          return new PreflightTx(params);
 
         default:
           throw new Error(`Unknown transaction for parsing: ${e.type}`);
@@ -917,19 +914,6 @@ export class TxRemoveBot extends EVMTx implements CMEvent {
   serialize(): TxSerialized {
     return {
       type: "TxAddBot",
-      content: JSON.stringify(this),
-    };
-  }
-}
-
-export class PreflightTx extends EVMTx {
-  toString(): string {
-    return "PreflightTx";
-  }
-
-  serialize(): TxSerialized {
-    return {
-      type: "PreflightTx",
       content: JSON.stringify(this),
     };
   }
