+++ src/index.ts
@@ -30,12 +30,12 @@ export * from "./payload/pool";
 export * from "./payload/token";
 export * from "./tokens/tokenData";
 export * from "./types";
+export { callRepeater } from "./utils/calls";
 export * from "./utils/errors";
 export * from "./utils/formatter";
 export * from "./utils/loading";
 export * from "./utils/math";
 export * from "./utils/price";
-export { callRepeater } from "./utils/repeater";
 export * from "./utils/types";
 export * from "./utils/validate";
 export * from "@gearbox-protocol/sdk-gov";
